# Conflicts:
#	.github/workflows/build.yml
#	.travis.yml
#	go.mod
#	go.sum
#	models/file.go
#	models/setting.go
#	models/user.go
#	pkg/filesystem/archive.go
#	pkg/filesystem/driver/local/handler.go
#	pkg/filesystem/driver/template/handler.go
#	pkg/filesystem/file.go
#	pkg/filesystem/filesystem.go
#	pkg/filesystem/hooks.go
#	pkg/filesystem/image.go
#	pkg/filesystem/manage.go
#	pkg/filesystem/upload.go
#	pkg/task/import.go
#	pkg/task/tranfer.go
#	service/share/manage.go

+++ pkg/task/tranfer.go
@@ -9,6 +9,7 @@ import (
 	"strings"
 
 	model "github.com/cloudreve/Cloudreve/v3/models"
+	"github.com/cloudreve/Cloudreve/v3/pkg/cluster"
 	"github.com/cloudreve/Cloudreve/v3/pkg/filesystem"
 	"github.com/cloudreve/Cloudreve/v3/pkg/filesystem/fsctx"
 	"github.com/cloudreve/Cloudreve/v3/pkg/util"
@@ -26,11 +27,14 @@ type TransferTask struct {
 
 // TransferProps 中转任务属性
 type TransferProps struct {
-	Src    []string `json:"src"`    // 原始文件
-	Parent string   `json:"parent"` // 父目录
-	Dst    string   `json:"dst"`    // 目的目录ID
+	Src      []string          `json:"src"`      // 原始文件
+	SrcSizes map[string]uint64 `json:"src_size"` // 原始文件的大小信息，从机转存时使用
+	Parent   string            `json:"parent"`   // 父目录
+	Dst      string            `json:"dst"`      // 目的目录ID
 	// 将会保留原始文件的目录结构，Src 除去 Parent 开头作为最终路径
 	TrimPath bool `json:"trim_path"`
+	// 负责处理中专任务的节点ID
+	NodeID uint `json:"node_id"`
 }
 
 // Props 获取任务属性
@@ -103,9 +107,9 @@ func (job *TransferTask) Do() {
 		fs.Tx = nil
 	}()
 
+	successCount := 0
 	for index, file := range job.TaskProps.Src {
 		_ = job.TaskModel.SetProgressTransaction(index, tx)
-
 		dst := path.Join(job.TaskProps.Dst, filepath.Base(file))
 		if job.TaskProps.TrimPath {
 			// 保留原始目录
@@ -114,10 +118,34 @@ func (job *TransferTask) Do() {
 			dst = path.Join(job.TaskProps.Dst, strings.TrimPrefix(src, trim))
 		}
 
-		ctx := context.WithValue(context.Background(), fsctx.DisableOverwrite, true)
-		err = fs.UploadFromPath(ctx, file, dst)
+		if job.TaskProps.NodeID > 1 {
+			// 指定为从机中转
+
+			// 获取从机节点
+			node := cluster.Default.GetNodeByID(job.TaskProps.NodeID)
+			if node == nil {
+				job.SetErrorMsg("从机节点不可用", nil)
+			}
+
+			// 切换为从机节点处理上传
+			fs.SwitchToSlaveHandler(node)
+			err = fs.UploadFromStream(context.Background(), &fsctx.FileStream{
+				File:        nil,
+				Size:        job.TaskProps.SrcSizes[file],
+				Name:        path.Base(dst),
+				VirtualPath: path.Dir(dst),
+				Src:         file,
+			}, false)
+		} else {
+			// 主机节点中转
+			err = fs.UploadFromPath(context.Background(), file, dst, 0)
+		}
+
 		if err != nil {
 			job.SetErrorMsg("文件转存失败", err)
+		} else {
+			successCount++
+			job.TaskModel.SetProgress(successCount)
 		}
 	}
 
@@ -147,15 +175,16 @@ func (job *TransferTask) Do() {
 
 // Recycle 回收临时文件
 func (job *TransferTask) Recycle() {
-	err := os.RemoveAll(job.TaskProps.Parent)
-	if err != nil {
-		util.Log().Warning("无法删除中转临时目录[%s], %s", job.TaskProps.Parent, err)
+	if job.TaskProps.NodeID == 1 {
+		err := os.RemoveAll(job.TaskProps.Parent)
+		if err != nil {
+			util.Log().Warning("无法删除中转临时目录[%s], %s", job.TaskProps.Parent, err)
+		}
 	}
-
 }
 
 // NewTransferTask 新建中转任务
-func NewTransferTask(user uint, src []string, dst, parent string, trim bool) (Job, error) {
+func NewTransferTask(user uint, src []string, dst, parent string, trim bool, node uint, sizes map[string]uint64) (Job, error) {
 	creator, err := model.GetActiveUserByID(user)
 	if err != nil {
 		return nil, err
@@ -168,6 +197,8 @@ func NewTransferTask(user uint, src []string, dst, parent string, trim bool) (Jo
 			Parent:   parent,
 			Dst:      dst,
 			TrimPath: trim,
+			NodeID:   node,
+			SrcSizes: sizes,
 		},
 	}
